requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-decouple
psycopg2-binary
pillow
drf-yasg
django-filter
drf-writable-nested    